+++ app/src/main/java/com/kojingu/xylo/codescreen/ListUsers.kt
@@ -1,0 +1,18 @@
+package com.kojingu.xylo.codescreen
+
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.sp
+import com.kojingu.xylo.R
+
+@Composable
+fun ListUsers(
+    userList: List<String>
+) {
+    for (names in userList) {
+        Text(
+            text = stringResource(id = R.string.list_users, names), fontSize = 20.sp
+        )
+    }
+}
